+++ source/objects/lander.lua
@@ -315,9 +315,14 @@ function Lander.create()
 	local lander = {}
 	lander.x = gintOriginX
 	lander.y = garrGround[lander.x] - 8
-	lander.sprite = garrImages[5]
-	lander.width = lander.sprite:getWidth()
-	lander.height = lander.sprite:getHeight()
+	--lander.sprite = garrImages[5]
+	--lander.width = lander.sprite:getWidth()
+	--lander.height = lander.sprite:getHeight()
+	
+	lander.spriteenum = enum.imageShip
+	lander.width = garrImages[lander.spriteenum]:getWidth()
+	lander.height = garrImages[lander.spriteenum]:getHeight()
+
 	-- 270 = up
 	lander.angle = 270
 	lander.vx = 0
@@ -446,9 +451,10 @@ function Lander.draw(worldOffset)
 			love.graphics.setColor(1,1,1,0.5)
 		end
 
-		-- TODO: for some reason ox = lander.width fails because lander.width is sometimes nil.
-		local ox = garrImages[5]:getWidth() / 2
-		local oy = garrImages[5]:getHeight() / 2
+		-- TODO: work out why lander.width doesn't work in mplayer mode
+		local ox = lander.width / 2
+		local oy = lander.height / 2
+		
 		love.graphics.draw(garrImages[5], drawingX,drawingY, math.rad(lander.angle), sx, sy, ox, oy)
 
 
@@ -460,7 +466,7 @@ function Lander.draw(worldOffset)
 			file and just use them here without the local keyword.
 		--]]
 		-- draw flames
-		local flameSprite	= garrImages[4]
+		local flameSprite	= garrImages[enum.imageFlameSprite]
 		local flameWidth	= flameSprite:getWidth()
 		local flameHeight	= flameSprite:getHeight()
 		local ox 			= flameWidth / 2